This fixes an issue where deduped node_modules can fail
The reason for the failure is as follows:

- we support optional peer dependency "mysql@1.2.3"
- we don't use mysql at all
- some other dependency we have has "mysql" as a dependency

This situation will create false positive signals that the
module exists (it exists, but the application doesn't need it to).

By checking the application's package.json for a mention of
"mysql", we check to see if this is really intended. If the
module is never mentioned, we consider it "not installed".